+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # История изменений
 
+## 1.0.7
+- Update linthtml-rules-htmlacademy
+- Adds htmlacademy rules
+- Remove bem for id
+- Remove `indent-style`
+- Remove `line-no-trailing-whitespace`
+- Remove `link-req-noopener`
+- Adds `spec-char-escape`
+- Fixed `head-meta-charset`
+
 ## 1.0.6
 - обновили linthtml-rules-htmlacademy 1.0.2
 - удалили `htmlacademy/attr-value-style`